wireless: ath9k: don't warn on -EPROBE_DEFER
This function gets called on probe, where -EPROBE_DEFER can trigger if
the GPIO driver is not ready.

Functions designed to deal with probe such as dev_err_probe do not
trigger on -EPROBE_DEFER. It doesn't make sense to call dev_err_probe
here as this function is not only called on probe.

Signed-off-by: Rosen Penev <rosenp@gmail.com>