net: ag71xx: use optional to get reset control
The code here is weird. There's an error check after the function which
is correct but a further one down does the opposite, which is always
true as it checks the opposite.

The intention seems to be to get an optional reset_control and check if
it's actually available. Although reset_control_(de)assert is NULL safe.

Signed-off-by: Rosen Penev <rosenp@gmail.com>