wifi: ath11k: switch to of_get_mac_address
This is needed to support nvmem defined MAC addresses in DTS.

In addition, check if the probe should be deferred as nvmem can load
after ath11k.

For brevity, ACPI is not a factor here. ath11k is too new for that.
Except for ath10k, all other users of device_get_mac_address are old
drivers with explicit ACPI support.

Signed-off-by: Rosen Penev <rosenp@gmail.com>